requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python-headless==4.9.0.80
numpy==1.24.3
onnxruntime==1.16.3
faiss-cpu==1.7.4
requests==2.31.0
watchdog==3.0.0

